add option to initiate first transmission
When running in a point to point configuration, if you need to
implement write-follow on both devices for flow control, the test
will hang on each end as both devices are waiting for the first
transmission, enabling the --initate-tx flag on one of the devices
will cause it to send the first transmission.

The condition is for the initate-tx flag to be set, and for rx_bytes
to equal tx_bytes. this is the case at the start of the test, when both
are zero. Additionally this sustains transmission, as in write-follow
only the number of bytes that have been received are transmitted back.
This means the number of rx bytes will never exceed the number of tx
bytes on the device initiating the transmission.